+++ src/redux/config/configStore.js
@@ -1,9 +1,10 @@
 import { configureStore } from "@reduxjs/toolkit";
-import user from "../modules/login/loginSlice";
+import user from "../modules/loginSlice";
 import orderdetail from "../modules/orderlistSlice";
 import menulist from "../modules/menulistSlice";
 import menu from "../modules/menuSlice";
 import restaurantlist from "../modules/restaurantlistSlice/restaurantSlice";
+
 const store = configureStore({
   reducer: { user, menulist, orderdetail, menu, restaurantlist },
 });
