--- src/redux/modules/login/setAuthorizationToken.js
@@ -1,10 +1,0 @@
-import axios from "axios";
-
-export default function setAuthorizationToken(token) {
-  if (token) {
-    axios.defaults.headers.common["Authorization"] = `${token}`;
-    console.log(token);
-  } else {
-    delete axios.defaults.headers.common["Authorization"];
-  }
-}
